+++ data/trx/ship/games/tr4/scripts/hall.lua
@@ -1,0 +1,9 @@
+require("tr4.inv_setup").apply({
+  key_item_2 = { scale = 1024, rot_x = -90 },
+  key_item_3 = { scale = 1024, rot_x = -90 },
+  puzzle_item_1 = { scale = 1280 },
+  puzzle_item_2 = { scale = 800, offset_y = 1 },
+  puzzle_item_3 = { scale = 800, offset_y = 1 },
+  puzzle_item_1_combo_1 = { scale = 384 },
+  puzzle_item_1_combo_2 = { scale = 1200 },
+})