+++ data/trx/ship/games/tr4/scripts/settomb1.lua
@@ -33,3 +33,10 @@ end)
 require("common.legend").setup(function()
   return trx.locale.get("general/legend")
 end)
+
+require("tr4.inv_setup").apply({
+  puzzle_item_1 = { scale = 1280, offset_y = 23 },
+  puzzle_item_2 = { scale = 768, offset_y = 24 },
+  puzzle_item_1_combo_1 = { scale = 1280, offset_y = 23 },
+  puzzle_item_1_combo_2 = { scale = 1280, offset_y = 23 },
+})